Add ScanCI function to RepositoryService in Bamboo
The code diff is primarily related to the addition of a new function, ScanCI, in the bamboo_repository_service go file. This function sends a POST request to trigger a CI scan for a specific repository in Bamboo, which enhances functionality of the Repository Service by providing a more hands-on control over the continuous integration process.