Update permission methods to include principal type
The permission management methods in the Confluence space permissions manager have been updated to include the principal type, either a user or a group. This change allows for more dynamic handling of permissions for different principal types. Also, updated the 'syncUser' and 'syncGroup' functions in the Jira actor lookup service to directly insert a user or a group instead of using variable references.